Create train_dataset
This repository includes a script that uses the PyCaret library to automate the training process of classification models. The script loads the main dataset (dataset.csv), sets the resultado column as the target variable, and performs a comparison among various algorithms to automatically select the most effective model. The selected model is then trained on the provided data, making it easy to reproduce and evaluate results in a simple and efficient way.