Create train_dataset
This repository includes code that uses the PyCaret library to automate the training process of classification models. The code loads the main dataset (dataset.csv), sets the resultado column as the target variable, and performs a comparison among various algorithms to automatically select the most effective model. The selected model is then trained on the provided data.