Create train_new_dataset
This repository includes code that uses the PyCaret library to automate the training process of classification models using a reduced dataset (new_dataset.csv). This dataset was derived from the original by removing irrelevant columns, keeping only the most significant features for the model. The code sets the resultado column as the target variable, compares several algorithms, and automatically selects the most effective model, which is then trained on the optimized data.