Create featureselection_featurewiz
This repository includes a script that uses the Featurewiz library to automate the feature selection process and trains a classification model using Logistic Regression on a preprocessed dataset (dataset.csv). This dataset was generated by removing irrelevant columns from the original, keeping only the most significant features for the model. The script defines the resultado column as the target variable, evaluates the model’s performance using the F1 score before and after feature selection, and selects the most effective model based on these results. The model is then retrained using only the selected features to optimize performance.